fix terrain in negative coordinates

+++ shared/math/perlin.h
@@ -1,5 +1,5 @@
 #ifndef __PERLIN_H
 #define __PERLIN_H
-float perlin_two(float x, float y, float gain, int octaves, int hgrid);
+float perlin_two(float x, float y, float gain, int octaves, float hgrid);
 void perlin_noise_shuf(void);
 #endif
